Refactors utilities

+++ top_mouse_tracker/utilities.py
@@ -5,18 +5,18 @@ Created on Mon Jan  7 17:56:59 2019
 
 @author: Thomas TOPILKO
 """
-
+import errno
 import os
+import sys
 
 import shutil
 import smtplib
 
-from TopMouseTracker import settings
-import TopMouseTracker.parameters as params
+import top_mouse_tracker.parameters as params
 
 
-def PrintColoredMessage(msg, color):
-    '''Prints a colored message in the console
+def print_colored_message(msg, color):
+    """Prints a colored message in the console
     
     Params : 
         msg (str) : the message to be printed
@@ -24,29 +24,29 @@ def PrintColoredMessage(msg, color):
         
     Returns :
         prints the colored message in the console
-    '''
+    """
     
     if isinstance(msg, str):
         if color in params.colors:
             print(params.colors[color]+msg+params.colors["off"])
         else:
-            PrintColoredMessage("[WARNING] The color {0} is not available!".format(color), "darkred")
+            print_colored_message("[WARNING] The color {0} is not available!".format(color), "darkred")
     else:
-        PrintColoredMessage("[WARNING] The message {0} is not in the correct format!".format(msg), "darkred")
+        print_colored_message("[WARNING] The message {0} is not in the correct format!".format(msg), "darkred")
 
 
-def PrintLoadingBar(percent):
-    '''Prints a loading bar in the console
-       e.g (0%) = '..................................................................................................................'
-           (11.5%) = '######............................................................................................................'
-           (100%) = '##################################################################################################################'
+def print_loading_bar(percent):  # FIXME: replace by tqdm
+    """Prints a loading bar in the console
+       e.g (0%) = '....................................................................................................'
+           (11.5%) = '######...........................................................................................'
+           (100%) = '##################################################################################################'
     
     Params :
         percent (int/float) : the % of the ongoing operation (e.g : 11.5 = 11.5%; 100 = 100%)
         
     Returns :
         prints the loading bar in the console
-    '''
+    """
     
     # Checks if percent argument is correct
     if percent >= 0:
@@ -61,33 +61,34 @@ def PrintLoadingBar(percent):
         return pattern
     
     else:
-        PrintColoredMessage("[WARNING] A progress cannot be negative right?", "darkred")
+        print_colored_message("[WARNING] A progress cannot be negative right?", "darkred")
 
 
-def HoursMinutesSeconds(time):
-    '''Function that takes a time variable (s) and converts it into (hh,mm,ss) format
+def hours_minutes_seconds(time):
+    """
+    Function that takes a time variable (s) and converts it into (hh,mm,ss) format
     
     Params : 
         time (int) : the number of seconds of a time/event
         
     Returns : 
         returns the time in (hh,mm,ss) format
-    '''
+    """
     # Checks if time argument is correct
     if time >= 0:
-        remainingMinutes = time % 3600
-        remainingSeconds = remainingMinutes % 60
+        remaining_minutes = time % 3600
+        remaining_seconds = remaining_minutes % 60
         hours = int(time/3600)
-        minutes = int(remainingMinutes/60)
-        seconds = int(remainingSeconds)
+        minutes = int(remaining_minutes/60)
+        seconds = int(remaining_seconds)
 
         return hours, minutes, seconds
     else:
-        PrintColoredMessage("[WARNING] Time cannot be negative right?","darkred")
+        print_colored_message("[WARNING] Time cannot be negative right?", "darkred")
 
 
-def PlaySound(n, sound):
-    '''Function that plays n times a sound
+def play_sound(n, sound):
+    """Function that plays n times a sound
     /!\ [WARNING] This function only works on OS X systems
     
     Params :
@@ -96,7 +97,7 @@ def PlaySound(n, sound):
         
     Output :
         Plays the sound "sound", "n" times from system speakers
-    '''
+    """
     # Checks if n argument is correct
     if n == 0 or n < 0:
         n = 1
@@ -104,22 +105,22 @@ def PlaySound(n, sound):
     # Checks if sound argument is correct
     if sound in params.sounds or sound is not None:  # TODO: check
         # Checks that the operating system is OS X
-        if settings._platform == "darwin":
+        if sys.platform == "darwin":
             for i in range(n):
                 os.system('afplay /System/Library/Sounds/'+str(sound)+'.aiff')
         else:
-            PrintColoredMessage("[WARNING] The functions utils.PlaySound requires OS X operating system to work!",
-                                "darkred")
+            print_colored_message("[WARNING] The functions utils.PlaySound requires OS X operating system to work!",
+                                  "darkred")
     else:
-        if settings._platform == "linux":
+        if sys.platform == "linux":
             for i in range(n):
                 os.system("/usr/bin/canberra-gtk-play --id='bell'")
         else:
-            PrintColoredMessage("[WARNING] The sound parameter for utils.PlaySound is unavailable or not correct!",
-                                "darkred")
+            print_colored_message("[WARNING] The sound parameter for utils.PlaySound is unavailable or not correct!",
+                                  "darkred")
 
 
-def CheckDirectoryExists(directory):
+def check_directory_exists(directory):
     ''' Function that checks is a certain directory exists. If not : creates it
     
     Params :
@@ -133,7 +134,7 @@ def CheckDirectoryExists(directory):
         if os.path.dirname(directory):
             if not os.path.exists(directory):
                 os.mkdir(directory)
-                PrintColoredMessage("[INFO] {0} has been created".format(directory), "darkgreen")
+                print_colored_message("[INFO] {0} has been created".format(directory), "darkgreen")
             else:
                 pass
         else:
@@ -142,52 +143,57 @@ def CheckDirectoryExists(directory):
         raise RuntimeError("{0} is not a valid directory!".format(directory))
 
 
-def ClearDirectory(directory):
+def clear_directory(directory):
     if isinstance(directory, str):
         if os.path.dirname(directory):
             if os.listdir(directory):
-                Input = input("Type [y]/n to clean {0} :   ".format(directory))
-                if Input == 'y':
+                _input = input("Type [y]/n to clean {0} :   ".format(directory))
+                if _input == 'y':
                     for obj in os.listdir(directory):
                         try:
                             os.remove(os.path.join(directory, obj))
-                        except:
-                            shutil.rmtree(os.path.join(directory, obj))
-                elif Input == 'n':
-                    PrintColoredMessage("{0} was kept untouched!".format(directory), "darkgreen")
+                        except OSError as err:
+                            if err.errno != errno.ENOENT:
+                                raise
+                            else:
+                                shutil.rmtree(os.path.join(directory, obj))
+
+                elif _input == 'n':
+                    print_colored_message("{0} was kept untouched!".format(directory), "darkgreen")
                 else:
-                    PrintColoredMessage("Wrong input!".format(directory), "darkgreen")
+                    print_colored_message("Wrong input!".format(directory), "darkgreen")
             else:
-                PrintColoredMessage("{0} was already empty!".format(directory), "darkgreen")
+                print_colored_message("{0} was already empty!".format(directory), "darkgreen")
         else:
             raise RuntimeError("{0} is not a valid directory!".format(directory))
     else:
         raise RuntimeError("{0} is not a valid directory!".format(directory))
 
 
-def CheckMail():
-    if params.mainParameters["email"] is not None:
-        params.mainParameters["password"] = input("Type the password for your email {0} : "
-                                                  .format(params.mainParameters["email"]))
+def check_mail():
+    if params.main["email"] is not None:
+        params.main["password"] = input("Type the password for your email {0} : "
+                                        .format(params.main["email"]))
         try:
-            s = smtplib.SMTP(params.mainParameters["smtp"], params.mainParameters["port"])
+            s = smtplib.SMTP(params.main["smtp"], params.main["port"])
             s.ehlo()
             s.starttls()
             s.ehlo()
-            s.login(params.mainParameters["email"], params.mainParameters["password"])
-            PrintColoredMessage("Emailing mode has been enabled", "darkgreen")
-        except:
-            PrintColoredMessage("[WARNING] Wrong Username or Password !", "darkred")
+            s.login(params.main["email"], params.main["password"])
+            print_colored_message("Emailing mode has been enabled", "darkgreen")
+        except smtplib.SMTPException as err:
+            print_colored_message("[WARNING] Wrong Username or Password !", "darkred")
+            print(err)
     else:
-        PrintColoredMessage("Emailing mode has been disabled", "darkgreen")
+        print_colored_message("Emailing mode has been disabled", "darkgreen")
 
 
-def CheckDirectories():
-    CheckDirectoryExists(params.mainParameters["tmtDir"])  # Checks if directory exists
-    CheckDirectoryExists(params.mainParameters["resultDir"])  # Checks if directory exists
-    ClearDirectory(params.mainParameters["resultDir"])  # Asks for clearing the directory if not empty
+def check_directories():
+    check_directory_exists(params.main["tmtDir"])  # Checks if directory exists
+    check_directory_exists(params.main["resultDir"])  # Checks if directory exists
+    clear_directory(params.main["resultDir"])  # Asks for clearing the directory if not empty
     
-    if params.savingParameters["fourcc"] == "Frames":
-        CheckDirectoryExists(params.mainParameters["segmentationDir"])  # Check if directory exists
+    if params.saving["fourcc"] == "Frames":
+        check_directory_exists(params.main["segmentationDir"])  # Check if directory exists
         
-    CheckMail()
+    check_mail()
